Refactor the assembly submodule
