--- src/felupe/_basis/__init__.py
@@ -1,3 +1,0 @@
-from ._basis import Basis, BasisMixed
-
-__all__ = ["Basis", "BasisMixed"]
