+++ src/felupe/assembly/_axi.py
@@ -20,68 +20,69 @@ import numpy as np
 
 from .._field._axi import FieldAxisymmetric
 from .._field._base import Field
-from ._weak import WeakForm
+from ._cartesian import IntegralFormCartesian
 
 
-class IntegralFormAxisymmetric(WeakForm):
+class IntegralFormAxisymmetric(IntegralFormCartesian):
     """An Integral Form for axisymmetric fields.
-    
-    Notes
-    -----
-    Axisymmetric scenarios are modeled with a 2D-mesh and consequently, a 2D element formulation. The rotation axis is chosen along the global X-axis :math:`(X,Y,Z) \widehat{=} (Z,R,\varphi)`. The 3x3 deformation gradient consists of an in-plane 2x2 sub-matrix and one additional entry for the out-of-plane stretch which is equal to the ratio of deformed and undeformed radius.
 
-..  math::
-    
-    \boldsymbol{F} = \begin{bmatrix} \boldsymbol{F}_{(2D)} & \boldsymbol{0} \\ \boldsymbol{0}^T & \frac{r}{R} \end{bmatrix}
-    
-    The variation of the deformation gradient consists of both in- and out-of-plane contributions.
-    
+        Notes
+        -----
+        Axisymmetric scenarios are modeled with a 2D-mesh and consequently, a 2D element formulation. The rotation axis is chosen along the global X-axis :math:`(X,Y,Z) \widehat{=} (Z,R,\varphi)`. The 3x3 deformation gradient consists of an in-plane 2x2 sub-matrix and one additional entry for the out-of-plane stretch which is equal to the ratio of deformed and undeformed radius.
+
     ..  math::
-        
-        \delta \boldsymbol{F}_{(2D)} = \delta \frac{\partial \boldsymbol{u}}{\partial \boldsymbol{X}} \qquad \text{and} \qquad \delta \left(\frac{r}{R}\right) = \frac{\delta u_r}{R}
-    
-    Again, the internal virtual work leads to two seperate terms.
-    
-    ..  math::
-        
-        -\delta W_{int} = \int_V \boldsymbol{P} : \delta \boldsymbol{F} \ dV = \int_V \boldsymbol{P}_{(2D)} : \delta \boldsymbol{F}_{(2D)} \ dV + \int_V \frac{P_{33}}{R} : \delta u_r \ dV
-    
-    The differential volume is further expressed as a product of the differential in-plane area and the differential arc length. The arc length integral is finally pre-evaluated.
-    
-    ..  math::
-    
-        \int_V dV = \int_{\varphi=0}^{2\pi} \int_A R\ dA\ d\varphi = 2\pi \int_A R\ dA
-    
-    Inserting the differential volume integral into the expression of internal virtual work, this leads to:
-    
-    ..  math::
-        
-        -\delta W_{int} = 2\pi \int_A \boldsymbol{P}_{(2D)} : \delta \boldsymbol{F}_{(2D)} \ R \ dA + 2\pi \int_A P_{33} : \delta u_r \ dA
-    
-    A Linearization of the internal virtual work expression gives four terms.
-    
-    ..  math::
-        
-        -\Delta \delta W_{int} &= \Delta_{(2D)} \delta_{(2D)} W_{int} + \Delta_{33} \delta_{(2D)} W_{int} + \Delta_{(2D)} \delta_{33} W_{int} + \Delta_{33} \delta_{33} W_{int}
-    
-        -\Delta_{(2D)} \delta_{(2D)} W_{int} &= 2\pi \int_A \delta \boldsymbol{F}_{(2D)} : \mathbb{A}_{(2D),(2D)} : \Delta \boldsymbol{F}_{(2D)} \ R \ dA
-    
-        -\Delta_{33} \delta_{(2D)} W_{int} &= 2\pi \int_A \delta \boldsymbol{F}_{(2D)} : \mathbb{A}_{(2D),33} : \Delta u_r \ dA
-    
-        -\Delta_{(2D)} \delta_{33} W_{int} &= 2\pi \int_A \delta u_r : \mathbb{A}_{33,(2D)} : \Delta \boldsymbol{F}_{(2D)} \ dA
-    
-        -\Delta_{33} \delta_{33} W_{int} &= 2\pi \int_A \delta u_r : \frac{\mathbb{A}_{33,33}}{R} : \Delta u_r \ dA
-    
-    with 
-    
-    ..  math::
-        
-        \mathbb{A}_{(2D),(2D)} &= \frac{\partial \psi}{\partial \boldsymbol{F}_{(2D)} \partial \boldsymbol{F}_{(2D)}}
-    
-        \mathbb{A}_{(2D),33} &= \frac{\partial \psi}{\partial \boldsymbol{F}_{(2D)} \partial F^3_{\hphantom{3}3}} \left ( = \mathbb{A}_{33,(2D)} \right )
-    
-        \mathbb{A}_{33,33} &= \frac{\partial \psi}{F^3_{\hphantom{3}3} \partial F^3_{\hphantom{3}3}}
+
+        \boldsymbol{F} = \begin{bmatrix} \boldsymbol{F}_{(2D)} & \boldsymbol{0} \\ \boldsymbol{0}^T & \frac{r}{R} \end{bmatrix}
+
+        The variation of the deformation gradient consists of both in- and out-of-plane contributions.
+
+        ..  math::
+
+            \delta \boldsymbol{F}_{(2D)} = \delta \frac{\partial \boldsymbol{u}}{\partial \boldsymbol{X}} \qquad \text{and} \qquad \delta \left(\frac{r}{R}\right) = \frac{\delta u_r}{R}
+
+        Again, the internal virtual work leads to two seperate terms.
+
+        ..  math::
+
+            -\delta W_{int} = \int_V \boldsymbol{P} : \delta \boldsymbol{F} \ dV = \int_V \boldsymbol{P}_{(2D)} : \delta \boldsymbol{F}_{(2D)} \ dV + \int_V \frac{P_{33}}{R} : \delta u_r \ dV
+
+        The differential volume is further expressed as a product of the differential in-plane area and the differential arc length. The arc length integral is finally pre-evaluated.
+
+        ..  math::
+
+            \int_V dV = \int_{\varphi=0}^{2\pi} \int_A R\ dA\ d\varphi = 2\pi \int_A R\ dA
+
+        Inserting the differential volume integral into the expression of internal virtual work, this leads to:
+
+        ..  math::
+
+            -\delta W_{int} = 2\pi \int_A \boldsymbol{P}_{(2D)} : \delta \boldsymbol{F}_{(2D)} \ R \ dA + 2\pi \int_A P_{33} : \delta u_r \ dA
+
+        A Linearization of the internal virtual work expression gives four terms.
+
+        ..  math::
+
+            -\Delta \delta W_{int} &= \Delta_{(2D)} \delta_{(2D)} W_{int} + \Delta_{33} \delta_{(2D)} W_{int} + \Delta_{(2D)} \delta_{33} W_{int} + \Delta_{33} \delta_{33} W_{int}
+
+            -\Delta_{(2D)} \delta_{(2D)} W_{int} &= 2\pi \int_A \delta \boldsymbol{F}_{(2D)} : \mathbb{A}_{(2D),(2D)} : \Delta \boldsymbol{F}_{(2D)} \ R \ dA
+
+            -\Delta_{33} \delta_{(2D)} W_{int} &= 2\pi \int_A \delta \boldsymbol{F}_{(2D)} : \mathbb{A}_{(2D),33} : \Delta u_r \ dA
+
+            -\Delta_{(2D)} \delta_{33} W_{int} &= 2\pi \int_A \delta u_r : \mathbb{A}_{33,(2D)} : \Delta \boldsymbol{F}_{(2D)} \ dA
+
+            -\Delta_{33} \delta_{33} W_{int} &= 2\pi \int_A \delta u_r : \frac{\mathbb{A}_{33,33}}{R} : \Delta u_r \ dA
+
+        with
+
+        ..  math::
+
+            \mathbb{A}_{(2D),(2D)} &= \frac{\partial \psi}{\partial \boldsymbol{F}_{(2D)} \partial \boldsymbol{F}_{(2D)}}
+
+            \mathbb{A}_{(2D),33} &= \frac{\partial \psi}{\partial \boldsymbol{F}_{(2D)} \partial F^3_{\hphantom{3}3}} \left ( = \mathbb{A}_{33,(2D)} \right )
+
+            \mathbb{A}_{33,33} &= \frac{\partial \psi}{F^3_{\hphantom{3}3} \partial F^3_{\hphantom{3}3}}
     """
+
     def __init__(self, fun, v, dV, u=None, grad_v=True, grad_u=True):
         R = v.radius
         self.dV = 2 * np.pi * R * dV
@@ -97,15 +98,15 @@ class IntegralFormAxisymmetric(WeakForm):
                     fun_2d = fun[:-1]
                     fun_zz = fun[-1].reshape(1, *fun[-1].shape) / R
 
-                form_a = WeakForm(fun_2d, v, self.dV, grad_v=grad_v)
-                form_b = WeakForm(fun_zz, v.scalar, self.dV)
+                form_a = IntegralFormCartesian(fun_2d, v, self.dV, grad_v=grad_v)
+                form_b = IntegralFormCartesian(fun_zz, v.scalar, self.dV)
 
                 self.forms = [form_a, form_b]
 
             else:
                 self.mode = 10
 
-                form_a = WeakForm(fun, v, self.dV, grad_v=False)
+                form_a = IntegralFormCartesian(fun, v, self.dV, grad_v=False)
                 self.forms = [
                     form_a,
                 ]
@@ -115,10 +116,10 @@ class IntegralFormAxisymmetric(WeakForm):
                 self.mode = 2
 
                 if grad_v and grad_u:
-                    form_aa = WeakForm(
+                    form_aa = IntegralFormCartesian(
                         fun[:-1, :-1, :-1, :-1], v, self.dV, u, True, True
                     )
-                    form_bb = WeakForm(
+                    form_bb = IntegralFormCartesian(
                         fun[-1, -1, -1, -1] / R**2,
                         v.scalar,
                         self.dV,
@@ -126,16 +127,18 @@ class IntegralFormAxisymmetric(WeakForm):
                         False,
                         False,
                     )
-                    form_ba = WeakForm(
+                    form_ba = IntegralFormCartesian(
                         fun[-1, -1, :-1, :-1] / R, v.scalar, self.dV, u, False, True
                     )
-                    form_ab = WeakForm(
+                    form_ab = IntegralFormCartesian(
                         fun[:-1, :-1, -1, -1] / R, v, self.dV, u.scalar, True, False
                     )
 
                 if not grad_v and grad_u:
-                    form_aa = WeakForm(fun[:-1, :-1, :-1], v, self.dV, u, False, True)
-                    form_bb = WeakForm(
+                    form_aa = IntegralFormCartesian(
+                        fun[:-1, :-1, :-1], v, self.dV, u, False, True
+                    )
+                    form_bb = IntegralFormCartesian(
                         fun[-1, -1, -1] / R**2,
                         v.scalar,
                         self.dV,
@@ -143,10 +146,10 @@ class IntegralFormAxisymmetric(WeakForm):
                         False,
                         False,
                     )
-                    form_ba = WeakForm(
+                    form_ba = IntegralFormCartesian(
                         fun[-1, :-1, :-1] / R, v.scalar, self.dV, u, False, True
                     )
-                    form_ab = WeakForm(
+                    form_ab = IntegralFormCartesian(
                         fun[:-1, -1, -1] / R, v, self.dV, u.scalar, False, False
                     )
 
@@ -155,15 +158,19 @@ class IntegralFormAxisymmetric(WeakForm):
             elif isinstance(v, FieldAxisymmetric) and isinstance(u, Field):
                 self.mode = 30
 
-                form_a = WeakForm(fun[:-1, :-1], v, self.dV, u, True, False)
-                form_b = WeakForm(fun[-1, -1] / R, v.scalar, self.dV, u, False, False)
+                form_a = IntegralFormCartesian(
+                    fun[:-1, :-1], v, self.dV, u, True, False
+                )
+                form_b = IntegralFormCartesian(
+                    fun[-1, -1] / R, v.scalar, self.dV, u, False, False
+                )
 
                 self.forms = [form_a, form_b]
 
             elif isinstance(v, Field) and isinstance(u, Field):
                 self.mode = 40
 
-                form_a = WeakForm(fun, v, self.dV, u, False, False)
+                form_a = IntegralFormCartesian(fun, v, self.dV, u, False, False)
 
                 self.forms = [
                     form_a,
